pyproject.toml
[tool.poetry]
name = "test-task-wg"
version = "0.1.0"
description = "wg_test_task"
authors = ["AleksandrKosmylev <alex.kosmylev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
psycopg2-binary = "^2.9.9"
flask-limiter = {extras = ["redis"], version = "^3.7.0"}
pytest = "^8.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
